+++ Runtime/Base/UnityEvents.cs
@@ -20,6 +20,7 @@ namespace PushForward.Base
 	[Serializable] public class StringEvent : UnityEvent<string> { }
 	[Serializable] public class IntEvent : UnityEvent<int> { }
 	[Serializable] public class FloatEvent : UnityEvent<float> { }
+	[Serializable] public class DoubleEvent : UnityEvent<double> { }
 	[Serializable] public class Vector2Event : UnityEvent<Vector2> { }
 	[Serializable] public class Vector3Event : UnityEvent<Vector3> { }
 	[Serializable] public class HitInfoEvent : UnityEvent<RaycastHit> { }