+++ Runtime/Timer.cs
@@ -32,7 +32,7 @@ namespace PushForward
 			[SerializeField] private TimeSpanEvent timeSpanEvent;
 
 			// "null value" time span
-			private TimeSpan triggerTime = new TimeSpan(-1);
+			private TimeSpan triggerTime = new(-1);
 
 			/// <summary>Check/set this event has been triggered already</summary>
 			public bool Triggered { get; set; }
@@ -43,7 +43,7 @@ namespace PushForward
 				{
 					if (this.triggerTime < TimeSpan.FromTicks(0))
 					{ this.triggerTime = TimeSpan.FromSeconds(this.triggerTimeInSeconds); }
-					
+
 					return this.triggerTime;
 				}
 			}
@@ -65,31 +65,32 @@ namespace PushForward
 			/// <param name="triggerTime">The time in which to trigger this event.</param>
 			/// <param name="actionToTrigger">The action to trigger.</param>
 			public static TimerEvent Create(TimeSpan triggerTime, TimeSpanEvent actionToTrigger)
-			{ return new TimerEvent { triggerTime = triggerTime, timeSpanEvent = actionToTrigger }; }
+				=> new() { triggerTime = triggerTime, timeSpanEvent = actionToTrigger };
 
 			/// <summary>Create a timer event to trigger at timer's end.</summary>
 			/// <remarks>While not setting a time, it is assumed this will trigger at the end.</remarks>
 			/// <param name="actionToTrigger">The action to trigger</param>
-			public static TimerEvent Create(TimeSpanEvent actionToTrigger)
-			{ return Create(TimeSpan.FromTicks(0), actionToTrigger); }
+			public static TimerEvent Create(TimeSpanEvent actionToTrigger) => Create(TimeSpan.FromTicks(0), actionToTrigger);
 		}
 		#endregion // timer action
 
 		#region Fields
 		#pragma warning disable IDE0044 // Add readonly modifier
 		/// <summary>The initial timer time.</summary>
-		[SerializeField] private double timerInSeconds;
+		[SerializeField] private double timerInSeconds = -1;
 		/// <summary>How to display the timer time.</summary>
 		[SerializeField] private DisplayMode displayMode;
-		/// <summary>Where to output the timer time.</summary>
-		[SerializeField] private UnityEngine.UI.Text outputText;
+		/// <summary>Reciever for timer output as double (in milliseconds).</summary>
+		[SerializeField] private DoubleEvent outputDouble;
+		/// <summary>Reciever for timer output as text.</summary>
+		[SerializeField] private StringEvent outputText;
 		/// <summary>The actions to take during the timer run.</summary>
 		[SerializeField] private TimerEvent[] timerActions;
 		/// <summary>The actions to take when the timer is over.</summary>
 		[SerializeField] private TimerEvent[] timerOverActions;
 		#pragma warning restore IDE0044 // Add readonly modifier
 
-		// the actual timer data
+		/// <summary>The actual timer data</summary>
 		private TimeSpan time;
 		#endregion // fields
 
@@ -117,42 +118,32 @@ namespace PushForward
 		#endregion // setup
 
 		#region Timer
-		/// <summary>Output the timer to the text component.</summary>
-		private void OutputToText()
+		private const string TimeStringDelimiter = ":";
+		private const string Time2Digits = "D2";
+		private const string Time3Digits = "D3";
+		/// <summary>Output the timer to the reciever.</summary>
+		private void OutputToRecievers()
 		{
+			this.outputDouble?.Invoke(this.time.TotalMilliseconds);
+
 			if (this.outputText == null)
 			{ return; }
-			
-			string output;
 
-			switch (this.displayMode)
-			{
-				case DisplayMode.HHMMSS:
-					output = this.time.Hours + ":"
-								+ this.time.Minutes.ToString("D2") + ":"
-								+ this.time.Seconds.ToString("D2");
-					break;
-				case DisplayMode.MMSS:
-					output = this.time.Minutes + ":" 
-					            + this.time.Seconds.ToString("D2");
-					break;
-				case DisplayMode.MMSStt:
-					output = this.time.Minutes + ":"
-					            + this.time.Seconds.ToString("D2") + ":"
-								+ this.time.Milliseconds.ToString("D3").Remove(2);
-					break;
-				case DisplayMode.SS: output = ((int)this.time.TotalSeconds).ToString(); break;
-				case DisplayMode.SStt: output = this.time.TotalSeconds.ToString("N3"); break;
-				// case DisplayMode.HHMMSStt:
-				default:
-					output = this.time.Hours + ":"
-					                         + this.time.Minutes.ToString("D2") + ":"
-					                         + this.time.Seconds.ToString("D2") + "."
-					                         + this.time.Milliseconds.ToString("D3").Remove(2);
-					break;
-			}
+			string output = this.displayMode switch
+				{
+					DisplayMode.HHMMSS => this.time.Hours + Timer.TimeStringDelimiter + this.time.Minutes.ToString(Timer.Time2Digits)
+										  + Timer.TimeStringDelimiter + this.time.Seconds.ToString(Timer.Time2Digits),
+					DisplayMode.MMSS => this.time.Minutes + Timer.TimeStringDelimiter + this.time.Seconds.ToString(Timer.Time2Digits),
+					DisplayMode.MMSStt => this.time.Minutes + Timer.TimeStringDelimiter + this.time.Seconds.ToString(Timer.Time2Digits)
+										  + Timer.TimeStringDelimiter + this.time.Milliseconds.ToString(Timer.Time3Digits).Remove(2),
+					DisplayMode.SS => ((int)this.time.TotalSeconds).ToString(),
+					DisplayMode.SStt => this.time.TotalSeconds.ToString("N3"),
+					_ => this.time.Hours + Timer.TimeStringDelimiter + this.time.Minutes.ToString(Timer.Time2Digits)
+						 + Timer.TimeStringDelimiter + this.time.Seconds.ToString(Timer.Time2Digits) + "." +
+						 this.time.Milliseconds.ToString(Timer.Time3Digits).Remove(2)
+				};
 
-			this.outputText.text = output;
+			this.outputText?.Invoke(output);
 		}
 
 		/// <summary>Timer Update method.</summary>
@@ -163,16 +154,14 @@ namespace PushForward
 			// make sure it doesn't dip below zero.
 			this.time = this.time.Max(0);
 
-			// show on text component
-			this.OutputToText();
+			this.OutputToRecievers();
 
 			// activate actions if available
-			this.timerActions?.DoForEach(
-				timerAction =>
-				{
-					if (!timerAction.Triggered && timerAction.TriggerTime >= this.time)
-					{ timerAction.Trigger(); }
-				});
+			this.timerActions?.DoForEach(timerAction =>
+										 {
+											 if (!timerAction.Triggered && timerAction.TriggerTime >= this.time)
+											 { timerAction.Trigger(); }
+										 });
 		}
 
 		/// <summary>Run all the actions that trigger at the end of the timer.</summary>
@@ -187,10 +176,11 @@ namespace PushForward
 		{
 			// define the predicate for when the timer should be running
 			bool TimerRunningPredicate() => this.time.Ticks > 0;
+
 			// run the timer update while predicate is true
-			this.ActionEachFrameWhilePredicate(this.UpdateTimer, TimerRunningPredicate);
+			this.ActionEachFrameWhilePredicate(TimerRunningPredicate, this.UpdateTimer);
 			// run the timer over actions when predicate is false
-			this.ActionWhenPredicate(this.TimerOverActions, () => !TimerRunningPredicate());
+			this.ActionWhenPredicate(() => !TimerRunningPredicate(), this.TimerOverActions);
 		}
 
 		/// <summary>Stop all timer functions. Time remaining is NOT reset.</summary>